requirements.txt
appdirs>=1.4.4
astropy>=6.0.0
pint-pulsar>=0.9.8
psrqpy>=1.2.8
tqdm>=4.66.2
ptmcmcsampler>=2.1.1
enterprise_extensions>=2.4.2
enterprise-pulsar @ git+https://github.com/M-Alakhras/enterprise
networkx>=3.21